requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tqdm==4.65.0
Pillow==10.0.0
google-generativeai>=0.4.0